more db change testing

+++ index.php
@@ -1,21 +1,30 @@
 <?php
 
- require_once('/var/config/printsheetmaker/config.php');
- $db_host = getenv('DB_HOST');
- echo "DB_HOST: $db_host <br>";
+ $path = '/var/config/printsheetmaker/config.php';
 
-try {
-    // Attempt to create a PDO instance and connect to the SQLite database
-    $db = new PDO("sqlite:$db_host");
+ if (file_exists($path)) {
+    require_once($path);
 
-    // Set PDO to throw exceptions on error
-    $db->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+    $db_host = getenv('DB_HOST');
+    echo "DB_HOST: $db_host <br>";
+   
+   try {
+       // Attempt to create a PDO instance and connect to the SQLite database
+       $db = new PDO("sqlite:$db_host");
+   
+       // Set PDO to throw exceptions on error
+       $db->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+   
+       // If successful, you can output a success message
+       echo "Connected to the SQLite database successfully.";
+   } catch (PDOException $e) {
+       // If an error occurs during connection, catch the exception and display an error message
+       echo "Connection failed: " . $e->getMessage();
+   }
 
-    // If successful, you can output a success message
-    echo "Connected to the SQLite database successfully.";
-} catch (PDOException $e) {
-    // If an error occurs during connection, catch the exception and display an error message
-    echo "Connection failed: " . $e->getMessage();
+
+} else {
+    echo "File does not exist at: $path";
 }
